+++ sdk/resourcemanager/compute/armcompute/zz_generated_constants.go
@@ -10,7 +10,7 @@ package armcompute
 
 const (
 	moduleName    = "armcompute"
-	moduleVersion = "v0.7.0"
+	moduleVersion = "v0.8.0"
 )
 
 type AccessLevel string
@@ -63,22 +63,6 @@ func PossibleArchitectureValues() []Architecture {
 	return []Architecture{
 		ArchitectureArm64,
 		ArchitectureX64,
-	}
-}
-
-// ArchitectureTypes - Specifies the Architecture Type
-type ArchitectureTypes string
-
-const (
-	ArchitectureTypesArm64 ArchitectureTypes = "Arm64"
-	ArchitectureTypesX64   ArchitectureTypes = "x64"
-)
-
-// PossibleArchitectureTypesValues returns the possible values for the ArchitectureTypes const type.
-func PossibleArchitectureTypesValues() []ArchitectureTypes {
-	return []ArchitectureTypes{
-		ArchitectureTypesArm64,
-		ArchitectureTypesX64,
 	}
 }
 
@@ -208,25 +192,6 @@ func PossibleConsistencyModeTypesValues() []ConsistencyModeTypes {
 		ConsistencyModeTypesApplicationConsistent,
 		ConsistencyModeTypesCrashConsistent,
 		ConsistencyModeTypesFileSystemConsistent,
-	}
-}
-
-// DataAccessAuthMode - Additional authentication requirements when exporting or uploading to a disk or snapshot.
-type DataAccessAuthMode string
-
-const (
-	// DataAccessAuthModeAzureActiveDirectory - When export/upload URL is used, the system checks if the user has an identity
-	// in Azure Active Directory and has necessary permissions to export/upload the data. Please refer to aka.ms/DisksAzureADAuth.
-	DataAccessAuthModeAzureActiveDirectory DataAccessAuthMode = "AzureActiveDirectory"
-	// DataAccessAuthModeNone - No additional authentication would be performed when accessing export/upload URL.
-	DataAccessAuthModeNone DataAccessAuthMode = "None"
-)
-
-// PossibleDataAccessAuthModeValues returns the possible values for the DataAccessAuthMode const type.
-func PossibleDataAccessAuthModeValues() []DataAccessAuthMode {
-	return []DataAccessAuthMode{
-		DataAccessAuthModeAzureActiveDirectory,
-		DataAccessAuthModeNone,
 	}
 }
 
@@ -962,6 +927,27 @@ func PossibleLinuxPatchAssessmentModeValues() []LinuxPatchAssessmentMode {
 	}
 }
 
+// LinuxVMGuestPatchAutomaticByPlatformRebootSetting - Specifies the reboot setting for all AutomaticByPlatform patch installation
+// operations.
+type LinuxVMGuestPatchAutomaticByPlatformRebootSetting string
+
+const (
+	LinuxVMGuestPatchAutomaticByPlatformRebootSettingAlways     LinuxVMGuestPatchAutomaticByPlatformRebootSetting = "Always"
+	LinuxVMGuestPatchAutomaticByPlatformRebootSettingIfRequired LinuxVMGuestPatchAutomaticByPlatformRebootSetting = "IfRequired"
+	LinuxVMGuestPatchAutomaticByPlatformRebootSettingNever      LinuxVMGuestPatchAutomaticByPlatformRebootSetting = "Never"
+	LinuxVMGuestPatchAutomaticByPlatformRebootSettingUnknown    LinuxVMGuestPatchAutomaticByPlatformRebootSetting = "Unknown"
+)
+
+// PossibleLinuxVMGuestPatchAutomaticByPlatformRebootSettingValues returns the possible values for the LinuxVMGuestPatchAutomaticByPlatformRebootSetting const type.
+func PossibleLinuxVMGuestPatchAutomaticByPlatformRebootSettingValues() []LinuxVMGuestPatchAutomaticByPlatformRebootSetting {
+	return []LinuxVMGuestPatchAutomaticByPlatformRebootSetting{
+		LinuxVMGuestPatchAutomaticByPlatformRebootSettingAlways,
+		LinuxVMGuestPatchAutomaticByPlatformRebootSettingIfRequired,
+		LinuxVMGuestPatchAutomaticByPlatformRebootSettingNever,
+		LinuxVMGuestPatchAutomaticByPlatformRebootSettingUnknown,
+	}
+}
+
 // LinuxVMGuestPatchMode - Specifies the mode of VM Guest Patching to IaaS virtual machine or virtual machines associated
 // to virtual machine scale set with OrchestrationMode as Flexible.
 // Possible values are:
@@ -1434,9 +1420,9 @@ func PossibleReplicationStatusTypesValues() []ReplicationStatusTypes {
 	}
 }
 
-// ResourceIdentityType - The type of identity used for the virtual machine. The type 'SystemAssigned, UserAssigned' includes
-// both an implicitly created identity and a set of user assigned identities. The type 'None' will
-// remove any identities from the virtual machine.
+// ResourceIdentityType - The type of identity used for the virtual machine scale set. The type 'SystemAssigned, UserAssigned'
+// includes both an implicitly created identity and a set of user assigned identities. The type 'None'
+// will remove any identities from the virtual machine scale set.
 type ResourceIdentityType string
 
 const (
@@ -2419,6 +2405,27 @@ func PossibleWindowsPatchAssessmentModeValues() []WindowsPatchAssessmentMode {
 	return []WindowsPatchAssessmentMode{
 		WindowsPatchAssessmentModeAutomaticByPlatform,
 		WindowsPatchAssessmentModeImageDefault,
+	}
+}
+
+// WindowsVMGuestPatchAutomaticByPlatformRebootSetting - Specifies the reboot setting for all AutomaticByPlatform patch installation
+// operations.
+type WindowsVMGuestPatchAutomaticByPlatformRebootSetting string
+
+const (
+	WindowsVMGuestPatchAutomaticByPlatformRebootSettingAlways     WindowsVMGuestPatchAutomaticByPlatformRebootSetting = "Always"
+	WindowsVMGuestPatchAutomaticByPlatformRebootSettingIfRequired WindowsVMGuestPatchAutomaticByPlatformRebootSetting = "IfRequired"
+	WindowsVMGuestPatchAutomaticByPlatformRebootSettingNever      WindowsVMGuestPatchAutomaticByPlatformRebootSetting = "Never"
+	WindowsVMGuestPatchAutomaticByPlatformRebootSettingUnknown    WindowsVMGuestPatchAutomaticByPlatformRebootSetting = "Unknown"
+)
+
+// PossibleWindowsVMGuestPatchAutomaticByPlatformRebootSettingValues returns the possible values for the WindowsVMGuestPatchAutomaticByPlatformRebootSetting const type.
+func PossibleWindowsVMGuestPatchAutomaticByPlatformRebootSettingValues() []WindowsVMGuestPatchAutomaticByPlatformRebootSetting {
+	return []WindowsVMGuestPatchAutomaticByPlatformRebootSetting{
+		WindowsVMGuestPatchAutomaticByPlatformRebootSettingAlways,
+		WindowsVMGuestPatchAutomaticByPlatformRebootSettingIfRequired,
+		WindowsVMGuestPatchAutomaticByPlatformRebootSettingNever,
+		WindowsVMGuestPatchAutomaticByPlatformRebootSettingUnknown,
 	}
 }