+++ sdk/resourcemanager/compute/armcompute/CHANGELOG.md
@@ -1,5 +1,46 @@
 # Release History
 
+## 0.8.0 (2022-04-18)
+### Breaking Changes
+
+- Const `DataAccessAuthModeAzureActiveDirectory` has been removed
+- Const `ArchitectureTypesX64` has been removed
+- Const `DataAccessAuthModeNone` has been removed
+- Const `ArchitectureTypesArm64` has been removed
+- Function `PossibleArchitectureTypesValues` has been removed
+- Function `PossibleDataAccessAuthModeValues` has been removed
+- Field `Architecture` of struct `SupportedCapabilities` has been removed
+- Field `DataAccessAuthMode` of struct `DiskProperties` has been removed
+- Field `DataAccessAuthMode` of struct `SnapshotProperties` has been removed
+- Field `DataAccessAuthMode` of struct `DiskUpdateProperties` has been removed
+- Field `DataAccessAuthMode` of struct `SnapshotUpdateProperties` has been removed
+- Field `Architecture` of struct `VirtualMachineImageProperties` has been removed
+
+### Features Added
+
+- New const `WindowsVMGuestPatchAutomaticByPlatformRebootSettingNever`
+- New const `LinuxVMGuestPatchAutomaticByPlatformRebootSettingUnknown`
+- New const `WindowsVMGuestPatchAutomaticByPlatformRebootSettingUnknown`
+- New const `LinuxVMGuestPatchAutomaticByPlatformRebootSettingIfRequired`
+- New const `LinuxVMGuestPatchAutomaticByPlatformRebootSettingNever`
+- New const `LinuxVMGuestPatchAutomaticByPlatformRebootSettingAlways`
+- New const `WindowsVMGuestPatchAutomaticByPlatformRebootSettingAlways`
+- New const `WindowsVMGuestPatchAutomaticByPlatformRebootSettingIfRequired`
+- New function `ProximityPlacementGroupPropertiesIntent.MarshalJSON() ([]byte, error)`
+- New function `PossibleWindowsVMGuestPatchAutomaticByPlatformRebootSettingValues() []WindowsVMGuestPatchAutomaticByPlatformRebootSetting`
+- New function `PossibleLinuxVMGuestPatchAutomaticByPlatformRebootSettingValues() []LinuxVMGuestPatchAutomaticByPlatformRebootSetting`
+- New struct `DedicatedHostGroupPropertiesAdditionalCapabilities`
+- New struct `LinuxVMGuestPatchAutomaticByPlatformSettings`
+- New struct `ProximityPlacementGroupPropertiesIntent`
+- New struct `WindowsVMGuestPatchAutomaticByPlatformSettings`
+- New field `UseRollingUpgradePolicy` in struct `AutomaticOSUpgradePolicy`
+- New field `AutomaticByPlatformSettings` in struct `LinuxPatchSettings`
+- New field `AdditionalCapabilities` in struct `DedicatedHostGroupProperties`
+- New field `Zones` in struct `ProximityPlacementGroup`
+- New field `AutomaticByPlatformSettings` in struct `PatchSettings`
+- New field `Intent` in struct `ProximityPlacementGroupProperties`
+
+
 ## 0.7.0 (2022-04-15)
 ### Breaking Changes